refactored netlify bug
