+++ src/components/Layout.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import Helmet from 'react-helmet'
 import { StaticQuery, graphql } from 'gatsby'
-import Header from './Header'
+import { Header } from '../components'
 import './layoutFont.css'
 
 const Layout = ({ children }) => (
