+++ src/components/index.js
@@ -2,7 +2,7 @@ export { default as Header } from './Header'
 export { default as HomePageHeroImage } from './HomePageHeroImage'
 export { default as HomePageTryptych } from './HomePageTryptych'
 export { default as Image } from './Image'
-export { default as Layout } from './layout'
+export { default as Layout } from './Layout'
 export { default as NavBar } from './NavBar'
 export { default as LayoutFont } from './LayoutFont'
 export { default as ProductPageContainer } from './ProductPageContainer'
